fix(backend): product_color_price/history created_by·approved_by 等列模型 i64→i32 对齐实库 INT4
- 同 m0044 fix_fk_types() 静默把 BIGINT→INTEGER 而模型仍 i64 的漂移(承 ai_* 修复同族):
  product_color_price(customer_id/created_by/approved_by)、color_price_history(operated_by/approved_by) i64→i32
- 相关 DTO(color_price_dto/history_dto)同步;handler 去 auth.user_id as i64、service 参数→i32、去 Some(.. as i32) workaround;迁移文本 BIGINT→INTEGER 自洽
- cargo fmt 0;只动 backend/。上报同族仍有 customer_color_price/custom_order/after_sales/process_node/log/sales_quotation 待一次性系统清

+++ backend/migration/src/domain/production/m0044_integrate_unreferenced_migrations.rs
@@ -739,7 +739,7 @@ ALTER TABLE "product_color_prices"
 
 -- 添加 customer_id 客户专属（NULL = 通用）
 ALTER TABLE "product_color_prices"
-    ADD COLUMN IF NOT EXISTS "customer_id" BIGINT REFERENCES "customers"("id");
+    ADD COLUMN IF NOT EXISTS "customer_id" INTEGER REFERENCES "customers"("id");
 
 -- 添加 season 季节标签
 ALTER TABLE "product_color_prices"
@@ -755,10 +755,10 @@ ALTER TABLE "product_color_prices"
 
 -- 添加创建人 / 审批人 / 审批时间
 ALTER TABLE "product_color_prices"
-    ADD COLUMN IF NOT EXISTS "created_by" BIGINT REFERENCES "users"("id");
+    ADD COLUMN IF NOT EXISTS "created_by" INTEGER REFERENCES "users"("id");
 
 ALTER TABLE "product_color_prices"
-    ADD COLUMN IF NOT EXISTS "approved_by" BIGINT REFERENCES "users"("id");
+    ADD COLUMN IF NOT EXISTS "approved_by" INTEGER REFERENCES "users"("id");
 
 ALTER TABLE "product_color_prices"
     ADD COLUMN IF NOT EXISTS "approved_at" TIMESTAMPTZ;
@@ -830,9 +830,9 @@ CREATE TABLE IF NOT EXISTS "color_price_history" (
     "change_reason" TEXT,
     "change_percent" DECIMAL(8,4),
     "quantity" DECIMAL(18,2),
-    "operated_by" BIGINT NOT NULL REFERENCES "users"("id"),
+    "operated_by" INTEGER NOT NULL REFERENCES "users"("id"),
     "operated_at" TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-    "approved_by" BIGINT REFERENCES "users"("id"),
+    "approved_by" INTEGER REFERENCES "users"("id"),
     "approved_at" TIMESTAMPTZ,
     "tenant_id" BIGINT NOT NULL,
     CONSTRAINT "chk_history_change_type" CHECK ("change_type" IN ('manual', 'batch', 'seasonal', 'customer_specific', 'tier'))
